rename: ProfileImageUtils -> AWSS3ImageUtils로 이름 변경
